+++ gibberish/detector/heuristics_detector/config.py
